Eliminate compiler warning.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@12129 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ cups/http-addr.c
@@ -871,7 +871,7 @@ httpGetHostname(http_t *http,		/* I - HTTP connection or NULL */
     */
 
     if (strlen(s) > 6 && !strcmp(s + strlen(s) - 6, ".local"))
-      strlcat(s, ".", slen);
+      strlcat(s, ".", (size_t)slen);
   }
 
  /*
